test: 💍 Add logout button test
