+++ src/components/logout-button/LogoutButton.tsx
@@ -1,20 +1,10 @@
 
 import { Button } from "@chakra-ui/react";
-import { useContext } from "react";
-import { UserType } from "../../layout/navigation-bar/authorized-navigation-bar/AuthorizedNavigationBar";
-import { UserContext } from "../../provider/UserProvider";
 
-export function LogoutButton() {
-    const { setUserDetail } = useContext(UserContext);
-    
-    const logout = () => {
-        console.log("logging out");
-        setUserDetail({
-            username: "",
-            role: UserType.GUEST,
-            userId: "",
-        });
-    }
+interface LogoutButtonProps {
+    logout: () => void,
+}
 
+export function LogoutButton({logout}: LogoutButtonProps) {
     return <Button onClick={logout} variant="navigationLink">Log out</Button>
 };
