+++ src/layout/navigation-bar/student-navigation-bar/StudentNavigationBar.tsx
@@ -2,8 +2,22 @@ import { RollCallNavigationBar } from "../roll-call-navigation-bar/RollCallNavig
 import NavigationBarLink from "../navigation-bar-link/NavigationBarLink";
 import { Box, Button, HStack } from "@chakra-ui/react";
 import { LogoutButton } from "../../../components/logout-button/LogoutButton";
+import { UserType } from "../authorized-navigation-bar/AuthorizedNavigationBar";
+import { useContext } from "react";
+import { UserContext } from "../../../provider/UserProvider";
 
 export function StudentNavigationBar() {
+  
+  const { setUserDetail } = useContext(UserContext);
+
+  const logout = () => {
+    setUserDetail({
+        username: "",
+        role: UserType.GUEST,
+        userId: "",
+    });
+  };
+
   return (
     <RollCallNavigationBar>
       <HStack>
@@ -21,7 +35,7 @@ export function StudentNavigationBar() {
       </HStack>
       <Box>
         <NavigationBarLink path="/">
-          <LogoutButton />
+          <LogoutButton logout={logout} />
         </NavigationBarLink>
       </Box>
     </RollCallNavigationBar>
