+++ src/layout/navigation-bar/teacher-navigation-bar/TeacherNavigationBar.tsx
@@ -2,8 +2,22 @@ import { RollCallNavigationBar } from "../roll-call-navigation-bar/RollCallNavig
 import NavigationBarLink from "../navigation-bar-link/NavigationBarLink";
 import { Box, Button, HStack } from "@chakra-ui/react";
 import { LogoutButton } from "../../../components/logout-button/LogoutButton";
+import { useContext } from "react";
+import { UserContext } from "../../../provider/UserProvider";
+import { UserType } from "../authorized-navigation-bar/AuthorizedNavigationBar";
 
 export function TeacherNavigationBar() {
+  
+  const { setUserDetail } = useContext(UserContext);
+
+  const logout = () => {
+    setUserDetail({
+        username: "",
+        role: UserType.GUEST,
+        userId: "",
+    });
+  };
+
   return (
     <RollCallNavigationBar>
       <HStack>
@@ -18,7 +32,7 @@ export function TeacherNavigationBar() {
       </HStack>
       <Box>
         <NavigationBarLink path="/">
-          <LogoutButton />
+          <LogoutButton logout={logout} />
         </NavigationBarLink>
       </Box>
     </RollCallNavigationBar>
